// src/app/model/peopleModel.ts
export class peopleModel {
    people_name: any;
    people_localtion_number: any;
    people_moo: any;
    people_road: any;
    people_alley: any;
    people_tumbon: any;
    people_district: any;
    people_province: any;
    people_postcode: any;
    people_phone: any;
    people_email: any;
    people_cardnumber: any;
    people_password: any;
    people_term: any;
    people_image_profile: any;
    people_generate: any;
    is_status: any; 
}